+++ example/ios/Runner/AppDelegate.swift
@@ -1,4 +1,5 @@
 import Adyen
+import adyen_checkout
 import Flutter
 import UIKit
 
@@ -15,5 +16,21 @@ import UIKit
     override func application(_: UIApplication, open url: URL, options _: [UIApplication.OpenURLOptionsKey: Any] = [:]) -> Bool {
         RedirectComponent.applicationDidOpen(from: url)
         return true
+    }
+}
+
+class AdyenAppearance: AdyenComponentAppearanceProvider, AdyenDropInAppearanceProvider {
+    static func createDropInStyle() -> Adyen.DropInComponent.Style {
+        var style = Adyen.DropInComponent.Style()
+        style.formComponent.mainButtonItem.button.backgroundColor = .black
+        style.formComponent.mainButtonItem.button.title.color = .white
+        return style
+    }
+
+    static func createCardComponentStyle() -> Adyen.FormComponentStyle {
+        var style = Adyen.FormComponentStyle()
+        style.mainButtonItem.button.backgroundColor = .black
+        style.mainButtonItem.button.title.color = .white
+        return style
     }
 }