+++ ios/Classes/components/delegates/ComponentActionHandler.swift
@@ -1,21 +1,21 @@
 import Adyen
 
-class ComponentActionDelegate: ActionComponentDelegate {
+class ComponentActionHandler: ActionComponentDelegate {
     private let componentFlutterApi: ComponentFlutterInterface
     private let componentId: String
-    private let finalizeAndDismissComponent: (Bool, @escaping (() -> Void)) -> Void
+    private let finalizeCallback: (Bool, @escaping (() -> Void)) -> Void
     
     init(
         componentFlutterApi: ComponentFlutterInterface,
         componentId: String,
-        finalizeAndDismissComponent: @escaping ((Bool, @escaping (() -> Void)) -> Void)
+        finalizeCallback: @escaping ((Bool, @escaping (() -> Void)) -> Void)
     ) {
         self.componentFlutterApi = componentFlutterApi
         self.componentId = componentId
-        self.finalizeAndDismissComponent = finalizeAndDismissComponent
+        self.finalizeCallback = finalizeCallback
     }
     
-    internal func didProvide(_ data: Adyen.ActionComponentData, from _: Adyen.ActionComponent) {
+    internal func didProvide(_ data: ActionComponentData, from _: ActionComponent) {
         do {
             let actionComponentData = ActionComponentDataModel(details: data.details.encodable, paymentData: data.paymentData)
             let actionComponentDataJson = try JSONEncoder().encode(actionComponentData)
@@ -30,34 +30,28 @@ class ComponentActionDelegate: ActionComponentDelegate {
                 completion: { _ in }
             )
         } catch {
-            finalizeAndDismissComponent(false) { [weak self] in
+            finalizeCallback(false) { [weak self] in
                 self?.sendErrorToFlutterLayer(error: error)
             }
         }
     }
 
-    internal func didComplete(from _: Adyen.ActionComponent) {
+    internal func didComplete(from _: ActionComponent) {
         // Only for voucher payment method - currently not supported.
     }
 
-    internal func didFail(with error: Error, from _: Adyen.ActionComponent) {
-        finalizeAndDismissComponent(false) { [weak self] in
+    internal func didFail(with error: Error, from _: ActionComponent) {
+        finalizeCallback(false) { [weak self] in
             self?.sendErrorToFlutterLayer(error: error)
         }
     }
     
     private func sendErrorToFlutterLayer(error: Error) {
-        let type: PaymentResultEnum
-        if let componentError = (error as? ComponentError), componentError == ComponentError.cancelled {
-            type = PaymentResultEnum.cancelledByUser
-        } else {
-            type = PaymentResultEnum.error
-        }
         let componentCommunicationModel = ComponentCommunicationModel(
             type: ComponentCommunicationType.result,
             componentId: componentId,
             paymentResult: PaymentResultDTO(
-                type: type,
+                type: .from(error: error),
                 reason: error.localizedDescription
             )
         )